requirements.txt
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
colorama==0.4.6
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.111.0
fastapi-cli==0.0.8
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
orjson==3.11.1
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pypng==0.20220715.0
python-dotenv==1.1.1
python-http-client==3.3.7
python-multipart==0.0.9
PyYAML==6.0.2
qrcode==7.4.2
rich==14.1.0
rich-toolkit==0.14.9
sendgrid==6.12.4
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.37.2
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
ujson==5.10.0
uvicorn==0.30.1
watchfiles==1.1.0
websockets==15.0.1
Werkzeug==3.1.3
